Add a toolchain system with "local" support
Add a toolchain system to make the relationship between rules_r and R
more explicit. The toolchain implemented in this commit is "local",
meaning that rules_r tries to find R within the system (using, e.g.,
`which R`). A specific version of R can be enforced, and on Mac OSX it
is possible to switch between various versions of R by specifying a
different "R_HOME".